requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
psycopg2-binary
gunicorn
scikit-learn
